Revert "(CR): i2c: qup: selective bus recovery"
This reverts commit 2198914bcdb98f9dd40b459215b51ab47c995a12.

The selective bus recovery does not work as planned.The recovery
does not happen even if it is really nedded. The check for the bus
error when a transfer timeout happened is not valid since the error
is cleared in interrupt handler.

Change-Id: I4abb9670ac0c91a2c3b8899a9179725842cc092d
Signed-off-by: Vladimir Karpovich <cvk011c@motorola.com>
Reviewed-on: http://gerrit.mot.com/650677
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Vladimir Karpovich <vkarpovich@motorola.com>
Reviewed-by: Lian-Wei Wang <lian-wei.wang@motorola.com>
Reviewed-by: Yi-Wei Zhao <gbjc64@motorola.com>
SLTApproved: Connie Zhao <czhao1@motorola.com>